Remove Index ABC, inline into FaissIndex
Index base class had exactly one subclass (FaissIndex) and LSHForest
never used it. All state management and save/load logic is now directly
in FaissIndex, removing ~220 lines of abstraction overhead.
