+++ src/tmap/index/faiss_index.py
@@ -17,13 +17,12 @@ from typing import Any
 import numpy as np
 from numpy.typing import NDArray
 
-from tmap.index.base import Index
 from tmap.index.types import KNNGraph
 
 _HNSW_THRESHOLD = 50_000
 
 
-class FaissIndex(Index):
+class FaissIndex:
     """Nearest neighbor index using FAISS.
 
     Parameters
@@ -50,9 +49,9 @@ class FaissIndex(Index):
         hnsw_ef_construction: int = 40,
         hnsw_ef_search: int = 64,
     ) -> None:
-        super().__init__(seed=seed)
         if mode not in {"auto", "flat", "hnsw"}:
             raise ValueError(f"mode must be auto/flat/hnsw, got {mode!r}")
+        self._seed = seed
         self._mode = mode
         self._hnsw_m = hnsw_m
         self._hnsw_ef_construction = hnsw_ef_construction
@@ -60,13 +59,53 @@ class FaissIndex(Index):
         self._effective_mode: str | None = None
         self._vectors: NDArray[np.float32] | None = None
         self._faiss_index = None
+        self._is_built = False
+        self._n_nodes: int = 0
+        self._metric: str | None = None
+
+    @property
+    def is_built(self) -> bool:
+        """Whether the index has been built and is ready for queries."""
+        return self._is_built
+
+    @property
+    def n_nodes(self) -> int:
+        """Number of points in the index."""
+        return self._n_nodes
+
+    @property
+    def metric(self) -> str | None:
+        """Distance metric used during build, or None if not yet built."""
+        return self._metric
 
     @property
     def effective_mode(self) -> str | None:
         """The actual index mode after building (None if not yet built)."""
         return self._effective_mode
 
-    def _build_from_vectors(self, vectors: NDArray[np.float32], metric: str) -> None:
+    def build_from_vectors(
+        self,
+        vectors: NDArray[np.float32],
+        metric: str = "euclidean",
+    ) -> FaissIndex:
+        """Build index from raw vectors.
+
+        Parameters
+        ----------
+        vectors : ndarray of shape (n_samples, n_features)
+            Data to index.
+        metric : str
+            ``"euclidean"`` or ``"cosine"``.
+
+        Returns
+        -------
+        self
+        """
+        if vectors.ndim != 2:
+            raise ValueError(f"vectors must be 2D, got shape {vectors.shape}")
+        if vectors.shape[0] < 2:
+            raise ValueError("Need at least 2 vectors to build index")
+
         import faiss
 
         vectors = np.ascontiguousarray(vectors, dtype=np.float32)
@@ -94,6 +133,140 @@ class FaissIndex(Index):
 
         self._faiss_index = index
         self._vectors = vectors
+        self._n_nodes = n
+        self._metric = metric
+        self._is_built = True
+        return self
+
+    def query_knn(self, k: int) -> KNNGraph:
+        """Get k-nearest neighbors for all points in the index.
+
+        Parameters
+        ----------
+        k : int
+            Number of neighbors per point.
+
+        Returns
+        -------
+        KNNGraph
+        """
+        self._check_is_built()
+        if k >= self._n_nodes:
+            raise ValueError(f"k={k} must be < n_nodes={self._n_nodes}")
+        return self._query_all(k)
+
+    def query_point(
+        self,
+        point: NDArray[np.float32],
+        k: int,
+    ) -> tuple[NDArray[np.int32], NDArray[np.float32]]:
+        """Query k-nearest neighbors for a single new point.
+
+        Returns
+        -------
+        (indices, distances)
+            Each of shape ``(k,)``.
+        """
+        self._check_is_built()
+        import faiss
+
+        query = np.ascontiguousarray(point.reshape(1, -1), dtype=np.float32)
+        if self._metric == "cosine":
+            faiss.normalize_L2(query)
+
+        distances, indices = self._faiss_index.search(query, k)
+
+        dists = distances[0].copy()
+        if self._metric == "cosine":
+            dists = 1.0 - dists
+        elif self._metric == "euclidean":
+            np.maximum(dists, 0, out=dists)
+            np.sqrt(dists, out=dists)
+
+        return indices[0].astype(np.int32), dists.astype(np.float32)
+
+    def query_batch(
+        self,
+        points: NDArray[np.float32],
+        k: int,
+    ) -> tuple[NDArray[np.int32], NDArray[np.float32]]:
+        """Query k-nearest neighbors for a batch of new points.
+
+        Returns
+        -------
+        (indices, distances)
+            Each of shape ``(m, k)``.
+        """
+        self._check_is_built()
+        import faiss
+
+        queries = np.ascontiguousarray(points, dtype=np.float32)
+        if self._metric == "cosine":
+            faiss.normalize_L2(queries)
+
+        distances, indices = self._faiss_index.search(queries, k)
+
+        dists = distances.copy()
+        if self._metric == "cosine":
+            dists = 1.0 - dists
+        elif self._metric == "euclidean":
+            np.maximum(dists, 0, out=dists)
+            np.sqrt(dists, out=dists)
+
+        return indices.astype(np.int32), dists.astype(np.float32)
+
+    def save(self, path: str | Path) -> None:
+        """Save index to disk."""
+        self._check_is_built()
+        import faiss
+
+        path = Path(path)
+        path.parent.mkdir(parents=True, exist_ok=True)
+
+        faiss.write_index(self._faiss_index, str(path))
+
+        meta = {
+            "n_nodes": self._n_nodes,
+            "seed": self._seed,
+            "metric": self._metric,
+            "mode": self._mode,
+            "effective_mode": self._effective_mode,
+            "hnsw_m": self._hnsw_m,
+            "hnsw_ef_construction": self._hnsw_ef_construction,
+            "hnsw_ef_search": self._hnsw_ef_search,
+        }
+        with open(str(path) + ".meta", "wb") as f:
+            pickle.dump(meta, f)
+
+    @classmethod
+    def load(cls, path: str | Path) -> FaissIndex:
+        """Load a previously saved index from disk."""
+        import faiss
+
+        path = Path(path)
+        with open(str(path) + ".meta", "rb") as f:
+            meta = pickle.load(f)
+
+        instance = cls(
+            seed=meta.get("seed"),
+            mode=meta.get("mode", "auto"),
+            hnsw_m=meta.get("hnsw_m", 32),
+            hnsw_ef_construction=meta.get("hnsw_ef_construction", 40),
+            hnsw_ef_search=meta.get("hnsw_ef_search", 64),
+        )
+        instance._faiss_index = faiss.read_index(str(path))
+        instance._n_nodes = meta.get("n_nodes", 0)
+        instance._metric = meta.get("metric")
+        instance._effective_mode = meta.get("effective_mode")
+        instance._is_built = True
+        instance._vectors = None
+        return instance
+
+    # -- internals --
+
+    def _check_is_built(self) -> None:
+        if not self._is_built:
+            raise RuntimeError("Index not built. Call build_from_vectors() first.")
 
     def _build_flat(self, d: int, metric_type: int) -> Any:
         import faiss
@@ -151,81 +324,3 @@ class FaissIndex(Index):
             indices.astype(np.int32),
             distances.astype(np.float32),
         )
-
-    def _query_single(
-        self,
-        point: NDArray[np.float32],
-        k: int,
-    ) -> tuple[NDArray[np.int32], NDArray[np.float32]]:
-        import faiss
-
-        query = np.ascontiguousarray(point.reshape(1, -1), dtype=np.float32)
-        if self._metric == "cosine":
-            faiss.normalize_L2(query)
-
-        distances, indices = self._faiss_index.search(query, k)
-
-        dists = distances[0].copy()
-        if self._metric == "cosine":
-            dists = 1.0 - dists
-        elif self._metric == "euclidean":
-            np.maximum(dists, 0, out=dists)
-            np.sqrt(dists, out=dists)
-
-        return indices[0].astype(np.int32), dists.astype(np.float32)
-
-    def _query_batch(
-        self,
-        points: NDArray[np.float32],
-        k: int,
-    ) -> tuple[NDArray[np.int32], NDArray[np.float32]]:
-        import faiss
-
-        queries = np.ascontiguousarray(points, dtype=np.float32)
-        if self._metric == "cosine":
-            faiss.normalize_L2(queries)
-
-        distances, indices = self._faiss_index.search(queries, k)
-
-        dists = distances.copy()
-        if self._metric == "cosine":
-            dists = 1.0 - dists
-        elif self._metric == "euclidean":
-            np.maximum(dists, 0, out=dists)
-            np.sqrt(dists, out=dists)
-
-        return indices.astype(np.int32), dists.astype(np.float32)
-
-    def _save_implementation(self, path: Path) -> None:
-        import faiss
-
-        faiss.write_index(self._faiss_index, str(path))
-
-        extra = {
-            "mode": self._mode,
-            "effective_mode": self._effective_mode,
-            "hnsw_m": self._hnsw_m,
-            "hnsw_ef_construction": self._hnsw_ef_construction,
-            "hnsw_ef_search": self._hnsw_ef_search,
-        }
-        with open(str(path) + ".faiss_meta", "wb") as f:
-            pickle.dump(extra, f)
-
-    def _load_implementation(self, path: Path) -> None:
-        import faiss
-
-        self._faiss_index = faiss.read_index(str(path))
-
-        meta_path = str(path) + ".faiss_meta"
-        try:
-            with open(meta_path, "rb") as f:
-                extra = pickle.load(f)
-            self._mode = extra.get("mode", "auto")
-            self._effective_mode = extra.get("effective_mode")
-            self._hnsw_m = extra.get("hnsw_m", 32)
-            self._hnsw_ef_construction = extra.get("hnsw_ef_construction", 40)
-            self._hnsw_ef_search = extra.get("hnsw_ef_search", 64)
-        except FileNotFoundError:
-            self._effective_mode = "flat"
-
-        self._vectors = None
